dom/media/test/test_eme_canvas_blocked.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test Encrypted Media Extensions</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
  <script type="text/javascript" src="manifest.js"></script>
  <script type="text/javascript" src="eme.js"></script>
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">
var manager = new MediaTestManager;

function startTest(test, token)
{
  manager.started(token);

  var sessions = [];

  var v = SetupEME(test, token);
  v.preload = "auto"; // Required due to "canplay" not firing for MSE unless we do this.

  v.addEventListener("loadeddata", function(ev) {
    var video = ev.target;
    var canvas = document.createElement("canvas");
    canvas.width = video.videoWidth;
    canvas.height = video.videoHeight;
    document.body.appendChild(canvas);
    var ctx = canvas.getContext("2d");
    var threwError = false;
    try {
      ctx.drawImage(video, 0, 0);
    } catch (ex) {
      threwError = true;
    }
    ok(threwError, TimeStamp(token) + " - Should throw an error when trying to draw EME video to canvas.");
    manager.finished(token);
  });

  LoadTestWithManagedLoadToken(test, v, manager, token,
                               { onlyLoadFirstFragments:2, noEndOfStream:true });
}

function beginTest() {
  manager.runTests(gEMETests, startTest);
}

SimpleTest.waitForExplicitFinish();
SetupEMEPref(beginTest);
</script>
</pre>
</body>
</html>
